+++ extensions/spine/src/spine/SkeletonRenderer.cpp
@@ -406,25 +406,28 @@ namespace spine {
 					triangles.indices = batch->allocateIndices(triangles.indexCount);
 					memcpy(triangles.indices, _clipper->getClippedTriangles().buffer(), sizeof(unsigned short) * _clipper->getClippedTriangles().size());
 
-					axmol::TrianglesCommand *batchedTriangles = batch->addCommand(renderer, _globalZOrder, texture, _programState, blendFunc, triangles, transform, transformFlags);
-
-					const float *verts = _clipper->getClippedVertices().buffer();
-					const float *uvs = _clipper->getClippedUVs().buffer();
-                    V3F_C4B_T2F *vertex = batchedTriangles->getTriangles().verts;
-                    for (int v = 0, vn = batchedTriangles->getTriangles().vertCount, vv = 0; v < vn; ++v, vv += 2, ++vertex) {
-                        vertex->vertices.x = verts[vv];
-                        vertex->vertices.y = verts[vv + 1];
+                    const float* verts  = _clipper->getClippedVertices().buffer();
+                    const float* uvs    = _clipper->getClippedUVs().buffer();
+                    V3F_C4B_T2F* vertex = triangles.verts;
+                    for (int v = 0, vn = triangles.vertCount, vv = 0; v < vn;
+                         ++v, vv += 2, ++vertex)
+                    {
+                        vertex->vertices.x  = verts[vv];
+                        vertex->vertices.y  = verts[vv + 1];
                         vertex->texCoords.u = uvs[vv];
                         vertex->texCoords.v = uvs[vv + 1];
-                        vertex->colors = color4B;
+                        vertex->colors      = color4B;
                     }
+					batch->addCommand(renderer, _globalZOrder, texture, _programState, blendFunc, triangles, transform, transformFlags);
 				} else {
 					// Not clipping.
-					axmol::TrianglesCommand *batchedTriangles = batch->addCommand(renderer, _globalZOrder, texture, _programState, blendFunc, triangles, transform, transformFlags);
-                    V3F_C4B_T2F *vertex = batchedTriangles->getTriangles().verts;
-                    for (int v = 0, vn = batchedTriangles->getTriangles().vertCount; v < vn; ++v, ++vertex) {
+                    V3F_C4B_T2F* vertex = triangles.verts;
+                    for (int v = 0, vn = triangles.vertCount; v < vn;
+                         ++v, ++vertex)
+                    {
                         vertex->colors = color4B;
                     }
+					batch->addCommand(renderer, _globalZOrder, texture, _programState, blendFunc, triangles, transform, transformFlags);
 				}
 			} else {
 				// Two color tinting.
@@ -444,28 +447,30 @@ namespace spine {
 					trianglesTwoColor.indices = twoColorBatch->allocateIndices(trianglesTwoColor.indexCount);
 					memcpy(trianglesTwoColor.indices, _clipper->getClippedTriangles().buffer(), sizeof(unsigned short) * _clipper->getClippedTriangles().size());
 
-					TwoColorTrianglesCommand *batchedTriangles = lastTwoColorTrianglesCommand = twoColorBatch->addCommand(renderer, _globalZOrder, texture, _programState, blendFunc, trianglesTwoColor, transform, transformFlags);
+                    const float* verts = _clipper->getClippedVertices().buffer();
+                    const float* uvs   = _clipper->getClippedUVs().buffer();
 
-					const float *verts = _clipper->getClippedVertices().buffer();
-					const float *uvs = _clipper->getClippedUVs().buffer();
-
-                    V3F_C4B_C4B_T2F *vertex = batchedTriangles->getTriangles().verts;
-                    for (int v = 0, vn = batchedTriangles->getTriangles().vertCount, vv = 0; v < vn; ++v, vv += 2, ++vertex) {
-                        vertex->position.x = verts[vv];
-                        vertex->position.y = verts[vv + 1];
+                    V3F_C4B_C4B_T2F* vertex = trianglesTwoColor.verts;
+                    for (int v = 0, vn = trianglesTwoColor.vertCount, vv = 0; v < vn;
+                         ++v, vv += 2, ++vertex)
+                    {
+                        vertex->position.x  = verts[vv];
+                        vertex->position.y  = verts[vv + 1];
                         vertex->texCoords.u = uvs[vv];
                         vertex->texCoords.v = uvs[vv + 1];
-                        vertex->color = color4B;
-                        vertex->color2 = darkColor4B;
+                        vertex->color       = color4B;
+                        vertex->color2      = darkColor4B;
                     }
+					lastTwoColorTrianglesCommand = twoColorBatch->addCommand(renderer, _globalZOrder, texture, _programState, blendFunc, trianglesTwoColor, transform, transformFlags);
 				} else {
-					TwoColorTrianglesCommand *batchedTriangles = lastTwoColorTrianglesCommand = twoColorBatch->addCommand(renderer, _globalZOrder, texture, _programState, blendFunc, trianglesTwoColor, transform, transformFlags);
-
-                    V3F_C4B_C4B_T2F *vertex = batchedTriangles->getTriangles().verts;
-                    for (int v = 0, vn = batchedTriangles->getTriangles().vertCount; v < vn; ++v, ++vertex) {
-                        vertex->color = color4B;
+                    V3F_C4B_C4B_T2F* vertex = trianglesTwoColor.verts;
+                    for (int v = 0, vn = trianglesTwoColor.vertCount; v < vn;
+                         ++v, ++vertex)
+                    {
+                        vertex->color  = color4B;
                         vertex->color2 = darkColor4B;
                     }
+					lastTwoColorTrianglesCommand = twoColorBatch->addCommand(renderer, _globalZOrder, texture, _programState, blendFunc, trianglesTwoColor, transform, transformFlags);
 				}
 			}
 			_clipper->clipEnd(*slot);