+++ templates/common/proj.win32/main.cpp
@@ -28,10 +28,18 @@
 #include "axmol.h"
 
 // Uncomment to enable win32 console
-// #define USE_WIN32_CONSOLE
+#define USE_WIN32_CONSOLE
 
 USING_NS_AX;
 
+int axmol_main() {
+    // create the application instance
+    AppDelegate app;
+    int ret = Application::getInstance()->run();
+    return ret;
+}
+
+#if !defined(_CONSOLE)
 int WINAPI _tWinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPTSTR lpCmdLine, int nCmdShow)
 {
     UNREFERENCED_PARAMETER(hPrevInstance);
@@ -43,8 +51,10 @@ int WINAPI _tWinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPTSTR lpCmdL
 #endif
 
     // create the application instance
-    AppDelegate app;
-    int ret = Application::getInstance()->run();
-
-    return ret;
+    return axmol_main();
 }
+#else
+int main(int, char**) {
+    return axmol_main();
+}
+#endif